% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\docType{data}
\name{alzheimer}
\alias{alzheimer}
\title{Real categorical data: alzheimer}
\description{
Presence or absence of 6 symptoms of Alzheimer's disease (AD) in 240 patients diagnosed with early onset AD conducted in the Mercer Institute in St. James's Hospital, Dublin.
A binary matrix, consisting of 240 rows and 6 columns, with each row denoting an individual and each column denoting the presence/absence of one of the 6 symptoms: Hallucination, Activity, Aggression, Agitation, Diurnal and Affective. A 1 denotes the presence of a symptom, a 0 the absence
}
\examples{
  data(alzheimer)
}
\references{
Moran M, Walsh C, Lynch A, Coen RF, Coakley D and Lawlor B (2004). Syndromes of behavioural and psychological symptoms in mild Alzheimer's disease. International Journal of Geriatric Psychiatry, 19(4), 359-364.

White A and Muprhy T (2014). BayesLCA: An R Pacakge for Bayesian Latent Class Analysis. Journal of Statistical Software, 61(13),1-28
}
\keyword{datasets}

