% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot.R
\docType{methods}
\name{barplot}
\alias{CoModesResults-method}
\alias{barplot}
\alias{barplot,}
\title{Barplot of a class [\code{\linkS4class{CoModesResults}}]}
\usage{
\S4method{barplot}{CoModesResults}(height, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}

\item{x}{an object of class [\code{\linkS4class{CoModesResults}}]}
}
\description{
Barplot of qualitative data from a [\code{\linkS4class{CoModesResults}}] object using parameters
to plot probablities of modalities.
}
\details{
Each line corresponds to one block of variable. Barplot is drawn for each cluster with the probabilities for 
each level (with is a mode for at least on cluster) to be in that cluster.
}
\seealso{
\code{\link{barplot}}
}

