\encoding{utf8}
\docType{data}
\name{birds}
\alias{birds}
\title{Birds data set}
\format{A data frame with 153 rows and 5 columns.}
\description{
Puffins are pelagic seabirds from the family Procellaridae. 
The data set of is composed by 153
puffins divided into three subspecies
dichrous (84 birds), lherminieri
(34 birds)
and subalaris (35 birds) is considered.
}
\arguments{
\item{collar}{1: none, ..., 5: continuous}
\item{eyebrows}{1: none, ..., 4: very pronounced}
\item{sub-caudal}{white, black, black and white, BLACK and white}
\item{border}{none, few, many}
\item{gender}{male,female}
}
\examples{
data("birds")
}
\references{
Bretagnolle,V.: 2007, Personal communication, source: Museum.
}
\keyword{datasets}
