% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\docType{package}
\name{CoModes-package}
\alias{CoModes}
\alias{CoModes-package}
\title{CoModes a package for clustering categorical data}
\description{
CoModes is a tool for clustering categorical data. The clustering goal is achieved by a mixture model which considers the
intra-class dependencies since the observed variables are grouped into conditionally independent blocks. The block distribution
is parsimonious since each block follows a multinomial distribution per modes. Under this distribution, the free parameters
correspond to the probabilities of the most probable levels (the modes) while the other levels (the non modes) are 
assumed to be equiprobable.
}
\details{
\tabular{ll}{
  Package: \tab CoModes\cr 
  Type: \tab Package\cr 
  Version: \tab 1.0.0\cr
  Date: \tab 2016-01-20\cr 
  License: \tab GPL-2\cr 
  LazyLoad: \tab yes\cr
}
}
\examples{
\dontrun{
a=1
}

}
\author{
Author: Marbac Matthieu, Biernacki Christophe and Vandewalle Vincent
}
\references{
Marbac Matthieu, Biernacki Christophe and Vandewalle Vincent (2015). Latent class model with conditional dependency per modes to cluster categorical data. arXiv:1402.5103
}
\keyword{package}

