\name{CoModes_criteria-class}
\alias{CoModes_criteria-class}
\title{The [\code{\linkS4class{CoModes_criteria}}] class}
\description{
  This is a class to contain results from CoModescluster function.
}
\details{
  This class is composed by tow attributs
  \describe{
    \item{loglike}{the log-likelihood value.}
    \item{bic}{the BIC criterion value.}
  } 
}
\examples{
  getSlots("CoModes_criteria")
}
