% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesCriteria-class}
\alias{CoModesCriteria-class}
\title{Constructor of [\code{\linkS4class{CoModesCriteria}}] class}
\description{
\describe{
  \item{loglike}{numeric. Log-likelihood.}
  \item{bic}{numeric. BIC criterion.}
}
}
\examples{
  getSlots("CoModesCriteria")

}

