% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesData-class}
\alias{CoModesData-class}
\title{Constructor of [\code{\linkS4class{CoModesData}}] class}
\description{
\describe{
  \item{data}{data.frame It contains the observations (rows correspond to individual and column to factor)}
  \item{modalities}{numeric. Number of levels for each categorical variables.}
  \item{levels}{list List of the levels for each categorical variables.}
}
}
\examples{
  getSlots("CoModesData")

}

