% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesIndiv-class}
\alias{CoModesIndiv-class}
\title{Constructor of [\code{\linkS4class{CoModesIndiv}}] class}
\description{
\describe{
  \item{proba}{matrix. Posterior probability per component.}
  \item{tik}{matrix. Fuzzy partition.}
  \item{partition}{numeric. Hard partition.}
}
}
\examples{
  getSlots("CoModesIndiv")

}

