% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesModel-class}
\alias{CoModesModel-class}
\title{Constructor of [\code{\linkS4class{CoModesModel}}] class}
\description{
\describe{
  \item{nbclasses}{numeric. Number of components.}
  \item{sigma}{numeric. Give the block where each variable is affiliated.}
  \item{modes}{mqtrix. Number of modes (rows correspond to components and columns to blocks)}
}
}
\examples{
  getSlots("CoModesModel")

}

