% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesParam-class}
\alias{CoModesParam-class}
\title{Constructor of [\code{\linkS4class{CoModesParam}}] class}
\description{
\describe{
  \item{proportions}{numeric. Component proportions.}
  \item{alpha}{list. Probabilities of the modes.}
}
}
\examples{
  getSlots("CoModesParam")

}

