% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesResults-class}
\alias{CoModesResults-class}
\title{Constructor of [\code{\linkS4class{CoModesResults}}] class}
\description{
This is a class of the results of CoModes.
}
\details{
\describe{
  \item{data}{}
  \item{model}{}
  \item{param}{}
  \item{indiv}{}
  \item{criteria}{}
}
}
\examples{
  getSlots("CoModesResults")

}

