% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\docType{data}
\name{lymphography}
\alias{lymphography}
\title{Real categorical data: lymphography}
\description{
This lymphography domain was obtained from the University Medical Centre, Institute of Oncology, Ljubljana, Yugoslavia.  Thanks go to M. Zwitter and M. Soklic for providing the data.
}
\details{
\itemize{
\item{x:}{Observed variables}
\item{class: }{normal find, metastases, malign lymph, fibrosis}
}

Informations are available here: \url{http://archive.ics.uci.edu/ml/datasets/Lymphography}
}
\examples{
  data(lymphography)
}
\keyword{datasets}

