% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\docType{data}
\name{seabirds}
\alias{seabirds}
\title{Real categorical data: seabirds}
\description{
The Seabirds data set is a biological data set describing 153 puffins (seabirds) by five plumage and external morphological characteristics. These seabirds are divided into three subspecies dichrous (84 birds), lherminieri (34 birds) and subalaris (35 birds).
}
\details{
\itemize{
\item{birds:}{plumage and external morphological characteristics}
\item{species: }{dichrous, lherminieri and subalaris}
}
}
\examples{
  data(seabirds)
}
\references{
Bretagnolle, V. (2007). Personal communication. source: Museum.
}
\keyword{datasets}

