% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\docType{data}
\name{acute}
\alias{acute}
\title{Real categorical data: acute}
\description{
The data was created by a medical expert as a data set to test the expert system, which will perform the presumptive diagnosis of two diseases of urinary system. The basis for rules detection was Rough Sets Theory. Each instance represents an potential patient. Here the first variable has been modified to only take integer value.
}
\details{
Informations are available here: http://archive.ics.uci.edu/ml/datasets/Acute+Inflammations

\itemize{
\item{acute[,1]: }{Temperature of patient (35C-42C)}
\item{acute[,2]: }{Occurrence of nausea (yes,no)}
\item{acute[,3]: }{Lumbar pain (yes,no)}
\item{acute[,4]: }{Urine pushing (continuous need for urination) (yes,no)}
\item{acute[,5]: }{Micturition pains (yes,no)}
\item{acute[,6]: }{Burning of urethra, itch, swelling of urethra outlet (yes,no)}
\item{urin: }{Inflammation of urinary bladder (yes,no)}
\item{renal: }{Nephritis of renal pelvis origin (yes,no)}
\item{urinrenal: }{Inflammation of urinary bladder  and Nephritis of renal pelvis origin (1:no-no, 2:yes-no, 3:no-yes, 4:yes-yes)} 
}
}
\examples{
  data(acute)
}
\references{
J.Czerniak, H.Zarzycki, Application of rough sets in the presumptive diagnosis of urinary system diseases, Artifical Inteligence and Security in Computing Systems, ACS'2002 9th International Conference Proceedings, Kluwer Academic Publishers,2003, pp. 41-51
}
\keyword{datasets}

