% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoModes.R
\name{CoModescluster}
\alias{CoModescluster}
\title{Create an instance of the [\code{\linkS4class{CoModesResults}}] class}
\usage{
CoModescluster(x, g, Gibbs_init = 30, Gibbs_iter = min(4000, (ncol(x) *
  400)), burnin = min(ncol(x) * 400, 4000), EM_init = 25,
  EM_tol = 10^(-3), nbcores = Gibbs_init)
}
\arguments{
\item{x}{data.frame, where each column is a factor.}

\item{g}{integer, defines the number of components.}

\item{Gibbs_init}{integer, number of chains performed for model selection (default 30).}

\item{Gibbs_iter}{integer, number of iterations of the MCMC algorithm for model selection (default min(ncol(x)*400,4000))}

\item{burnin}{integer, number of iterations of the burn-in of the MCMC algorithm (default min(ncol(x)*400,4000))}

\item{EM_init}{integer, number of runs of EM algorithm for parameter inference (default 25)}

\item{EM_tol}{numeric, tolerance for the stopping criterion of the EM algorithm (default 0.001)}

\item{nbcores}{number of cores used by the algorithm (only for Linux and MAC). (default Gibbs_init)}
}
\value{
Returns an instance of the [\code{\linkS4class{CoModesResults}}] class.
}
\description{
This function performs the model selection and the parameter inference.
x, g, Gibbs_init=30,  Gibbs_iter=min(4000,(ncol(x)*400)), burnin=min(ncol(x)*400,4000), EM_init=25, EM_tol=10^(-3), nbcores=Gibbs_init
}
\examples{
\dontrun{
ess
}

}

