% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{CoModesResults-class}
\alias{CoModesResults-class}
\title{Constructor of [\code{\linkS4class{CoModesResults}}] class}
\description{
This S4 class contains the results from the function \link{CoModescluster}.
 
\describe{
  \item{data}{CoModesData. Information about the data.}
  \item{model}{CoModesModel. Information about the model.}
  \item{param}{CoModesParam. Information about the parameters.}
  \item{indiv}{CoModesIndiv. Information about the individuals.}
  \item{criteria}{CoModesCriteria. Information about the criteria.}
}
}
\examples{
  getSlots("CoModesResults")

}

