% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,CoModesResults-method}
\title{Summary of a class [\code{\linkS4class{CoModesResults}}]}
\usage{
\S4method{summary}{CoModesResults}(object)
}
\arguments{
\item{object}{an object of class [\code{\linkS4class{CoModesResults}}]}
}
\description{
Summary of a class [\code{\linkS4class{CoModesResults}}]
}

